Cleanup: widen func_ovl2_800EBC0C arg0 from s32 to FTStruct *
Callers in mpcommon.c pass FTStruct * through an (s32)(intptr_t) cast
added in the flag-5 pass. Body doesn't read arg0, so the warning was
silenced rather than fixed — the prototype still truncated to 32 bits
on LP64 if anyone ever used the arg.

Widen the prototype to FTStruct * (matches every caller) and drop the
casts. Follow-up 4 from docs/cmake_warning_audit_handoff_2026-04-20.md.

+++ src/mp/mpcommon.c
@@ -376,7 +376,7 @@ void mpCommonUpdateFighterSlopeContour(GObj *fighter_gobj)
             {
                 joint = fp->joints[attr->joint_rfoot_id];
 
-                func_ovl2_800EBC0C((s32)(intptr_t)fp, &sp30, &sp2C, attr->joint_rfoot_rotate, joint);
+                func_ovl2_800EBC0C(fp, &sp30, &sp2C, attr->joint_rfoot_rotate, joint);
 
                 if (func_ovl2_800DDF74(fighter_gobj, fp, attr, joint, &sp30) != FALSE)
                 {
@@ -388,7 +388,7 @@ void mpCommonUpdateFighterSlopeContour(GObj *fighter_gobj)
             {
                 joint = fp->joints[attr->joint_lfoot_id];
 
-                func_ovl2_800EBC0C((s32)(intptr_t)fp, &sp30, &sp2C, attr->joint_lfoot_rotate, joint);
+                func_ovl2_800EBC0C(fp, &sp30, &sp2C, attr->joint_lfoot_rotate, joint);
 
                 if (func_ovl2_800DDF74(fighter_gobj, fp, attr, joint, &sp30) != FALSE)
                 {
